gave triangular, circular and rectangular orbits the capability to choose orbit location. Removed clippiness at beginning of orbit (circle is not perfect)

+++ src/com/jtronlabs/enemy_types_orbiters/Orbiter_Circle_IncSpeedOnHitView.java
@@ -11,7 +11,7 @@ public class Orbiter_Circle_IncSpeedOnHitView extends Orbiter_CircleView impleme
 	public Orbiter_Circle_IncSpeedOnHitView(Context context,int score,double speedY, double speedX,double collisionDamage, 
 			double health, double bulletFreq,
 			float heightView,float widthView,double bulletDamage,double bulletVerticalSpeed,double probSpawnBeneficialObjecyUponDeath,
-			double circleRadius,int angularVelocityInDegrees) {
+			int circleRadius,int angularVelocityInDegrees) {
 		super( context, score, speedY,  speedX, collisionDamage, 
 				 health,  bulletFreq,
 				 heightView, widthView, bulletDamage, bulletVerticalSpeed, probSpawnBeneficialObjecyUponDeath,
